+++ vogs/src/swap_data.c
@@ -12,12 +12,12 @@
 
 #include "push_swap.h"
 
-void			swap_data(t_block **head)
+void				swap_data(t_block **head)
 {
-	long long	swap;
-	int		len;
-	t_block	*temp1;
-	t_block	*temp2;
+	long 			swap;
+	unsigned int	len;
+	t_block			*temp1;
+	t_block			*temp2;
 
 	len = list_length((*head));
 	if (len < 1)
